feat(global): move navigation select to burger menu
